services/engine/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
supabase>=2.3.0
httpx>=0.26.0
python-dotenv>=1.0.0
structlog>=24.1.0
arq>=0.25.0
xmltodict>=0.13.0
tenacity>=8.2.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-env>=1.1.0
